refactor(components): rename dashboard.tsx to Dashboard.tsx
Resolves the lowercase-vs-PascalCase overlap in `src/components/`:
the file `dashboard.tsx` was the only top-level component (outside
`src/components/ui/`, which intentionally follows the shadcn kebab-case
convention) using a non-PascalCase name, and it sat alongside a
`dashboard/` folder of related sub-components, which is confusing on
case-insensitive filesystems and to a reader scanning imports.

Renames the file via `git mv` to preserve history, and updates the only
consumer (`src/app/dashboard/page.tsx`) to import from
`@/components/Dashboard`. The `dashboard/` folder of children stays put
since its contents are addressed under their own paths
(`@/components/dashboard/DocumentGrid`, etc.) and don't collide.

Verification: 76/76 Jest tests pass, `tsc --noEmit` reports no new
errors (only the pre-existing jest-dom matcher typing issue in
component tests, unrelated to this rename).

Refs: p2-23-naming
