+++ src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import { Dashboard } from "@/components/dashboard";
+import { Dashboard } from "@/components/Dashboard";
 
 const Page = async () => {
   // Middleware already protects this route
